Fixed some bugs with files. Making code to look clenaer. Getting ready to send it in.

+++ src/Managers/FileManager.java
@@ -18,12 +18,26 @@ public final class FileManager {
         for (Film f : hraneFilmy){
             str = f.getName() + ".txt";
             try {
-                FileWriter fw = new FileWriter(new File(dirHrane, str), true);
-                fw.write(f.getName()+";");
-                fw.write(f.getDirector()+";");
-                fw.write(f.getRokVydani()+";");
-                fw.write(f.getStaff()+";");
-                fw.close();
+                File file = new File(dirHrane, str);
+                if (file.exists() && !f.isChanged()){
+                    System.out.println("Film: " + f.getName() +  " v souborech už mám.");
+                }else if (file.exists() && f.isChanged()){
+                    file.delete();
+                    FileWriter fw = new FileWriter(new File(dirHrane, str), true);
+                    fw.write(f.getName()+";");
+                    fw.write(f.getDirector()+";");
+                    fw.write(f.getRokVydani()+";");
+                    fw.write(f.getStaff()+";");
+                    fw.close();
+                }else {
+                    file.delete();
+                    FileWriter fw = new FileWriter(new File(dirHrane, str), true);
+                    fw.write(f.getName()+";");
+                    fw.write(f.getDirector()+";");
+                    fw.write(f.getRokVydani()+";");
+                    fw.write(f.getStaff()+";");
+                    fw.close();
+                }
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
@@ -31,13 +45,27 @@ public final class FileManager {
         for (FilmAnimated f : animovaneFilmy){
             str = f.getName() + ".txt";
             try {
-                FileWriter fw = new FileWriter(new File(dirAnimovane, str), true);
-                fw.write(f.getName()+";");
-                fw.write(f.getDirector()+";");
-                fw.write(f.getRokVydani()+";");
-                fw.write(f.getMinVek()+";");
-                fw.write(f.getStaff()+";");
-                fw.close();
+                File file = new File(dirAnimovane, str);
+                if (file.exists() && !f.isChanged()){
+                    System.out.println("Animák: " + f.getName() +  " v souborech už mám.");
+                }else if (file.exists() && f.isChanged()){
+                    file.delete();
+                    FileWriter fw = new FileWriter(new File(dirAnimovane, str), true);
+                    fw.write(f.getName()+";");
+                    fw.write(f.getDirector()+";");
+                    fw.write(f.getRokVydani()+";");
+                    fw.write(f.getMinVek()+";");
+                    fw.write(f.getStaff()+";");
+                    fw.close();
+                }else {
+                    FileWriter fw = new FileWriter(new File(dirAnimovane, str), true);
+                    fw.write(f.getName()+";");
+                    fw.write(f.getDirector()+";");
+                    fw.write(f.getRokVydani()+";");
+                    fw.write(f.getMinVek()+";");
+                    fw.write(f.getStaff()+";");
+                    fw.close();
+                }
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
@@ -48,10 +76,8 @@ public final class FileManager {
         String dirHrane = aktDir + File.separator + "src" + File.separator + "MovieFiles/Hrane";
         String dirAnimovane = aktDir + File.separator + "src" + File.separator + "MovieFiles/Animovane";
 
-        String str;
-
-        List<String> hraneFilmySoubory = new ArrayList<String>();
-        List<String> animovaneFilmySoubory = new ArrayList<String>();
+        List<String> hraneFilmySoubory = new ArrayList<>();
+        List<String> animovaneFilmySoubory = new ArrayList<>();
 
         File[] hraneFolder = new File(dirHrane).listFiles();
         File[] animovaneFolder = new File(dirAnimovane).listFiles();
@@ -70,7 +96,7 @@ public final class FileManager {
         hraneFilmySoubory.forEach((value)->{
             File vstupniSoubor = new File(dirHrane + "/" + value);
             try {
-                String name, director, staff;
+                String name, director;
                 Integer rokVydani;
                 List<String> staffToBeAdded = new ArrayList<>();
                 FileReader fr = new FileReader(vstupniSoubor);
@@ -81,15 +107,11 @@ public final class FileManager {
                     name = splited[0];
                     director = splited[1];
                     rokVydani = Integer.valueOf(splited[2]);
-                    String[] staffSplit = (splited[3].substring(1, splited[3].length() - 1)).split(", "); //removes brackets, split by comma and put it in array
+                    String[] staffSplit = (splited[3].substring(1, splited[3].length() - 1)).split(", ");
                     for (String s : staffSplit){
                         staffToBeAdded.add(s);
                     }
-                    Film film = new Film(name, director, rokVydani);
-                    hraneFilmy.add(film);
-                    for (String actor : staffToBeAdded){
-                        film.addStaff(actor);
-                    }
+                    AddFilm(hraneFilmy, name, director, rokVydani, staffToBeAdded);
                 }
             } catch (FileNotFoundException e) {
                 throw new RuntimeException(e);
@@ -102,7 +124,7 @@ public final class FileManager {
         animovaneFilmySoubory.forEach((value)->{
             File vstupniSoubor = new File(dirAnimovane + "/" + value);
             try {
-                String name, director, staff;
+                String name, director;
                 Integer rokVydani, doporucenyVek;
                 List<String> staffToBeAdded = new ArrayList<>();
                 FileReader fr = new FileReader(vstupniSoubor);
@@ -114,15 +136,11 @@ public final class FileManager {
                     director = splited[1];
                     rokVydani = Integer.valueOf(splited[2]);
                     doporucenyVek = Integer.valueOf(splited[3]);
-                    String[] staffSplit = (splited[4].substring(1, splited[4].length() - 1)).split(", "); //removes brackets, split by comma and put it in array
+                    String[] staffSplit = (splited[4].substring(1, splited[4].length() - 1)).split(", ");
                     for (String s : staffSplit){
                         staffToBeAdded.add(s);
                     }
-                    FilmAnimated film = new FilmAnimated(name, director, rokVydani, doporucenyVek);
-                    animovaneFilmy.add(film);
-                    for (String actor : staffToBeAdded){
-                        film.addStaff(actor);
-                    }
+                    AddAnimated(animovaneFilmy, name, director, rokVydani, doporucenyVek, staffToBeAdded);
                 }
             } catch (FileNotFoundException e) {
                 throw new RuntimeException(e);
@@ -131,5 +149,44 @@ public final class FileManager {
                 System.out.println(e);
             }
         });
+    }
+
+    public static void AddAnimated(List<FilmAnimated> animovaneFilmy, String name, String director, Integer rokVydani, Integer doporucenyVek, List<String> staffToBeAdded){
+        ArrayList<FilmAnimated> tobeAdded = new ArrayList<>();
+        boolean flag = false;
+        for (FilmAnimated f : animovaneFilmy){
+            if (f.getName().equals(name)){
+                System.out.println("Animák s jménem: " + name + " už mám v programu načtený.");
+                flag = true;
+            } else {
+                FilmAnimated film = new FilmAnimated(name, director, rokVydani, doporucenyVek);
+                tobeAdded.add(film);
+                for (String actor : staffToBeAdded){
+                    film.addStaff(actor);
+                }
+            }
+        }
+        if (!flag){
+            animovaneFilmy.add(tobeAdded.get(0));
+        }
+    }
+    public static void AddFilm(List<Film> hraneFilmy, String name, String director, Integer rokVydani, List<String> staffToBeAdded){
+        ArrayList<Film> tobeAdded = new ArrayList<>();
+        boolean flag = false;
+        for (Film f : hraneFilmy){
+            if (f.getName().equals(name)){
+                System.out.println("Film s jménem: " + name + " už mám v programu načtený.");
+                flag = true;
+            } else {
+                Film film = new Film(name, director, rokVydani);
+                tobeAdded.add(film);
+                for (String actor : staffToBeAdded){
+                    film.addStaff(actor);
+                }
+            }
+        }
+        if (!flag){
+            hraneFilmy.add(tobeAdded.get(0));
+        }
     }
 }
